+++ lib/Sema/TypeCheckType.cpp
@@ -182,58 +182,6 @@ Type TypeChecker::getExceptionType(DeclContext *dc, SourceLoc loc) {
   return Type();
 }
 
-static Type getObjectiveCNominalType(TypeChecker &TC,
-                                     Type &cache,
-                                     Identifier ModuleName,
-                                     Identifier TypeName,
-                                     DeclContext *dc) {
-  if (cache)
-    return cache;
-
-  auto &Context = TC.Context;
-
-  // FIXME: Does not respect visibility of the module.
-  ModuleDecl *module = Context.getLoadedModule(ModuleName);
-  if (!module)
-    return nullptr;
-
-  NameLookupOptions lookupOptions
-    = defaultMemberLookupOptions |
-      NameLookupFlags::KnownPrivate;
-  if (auto result = TC.lookupMemberType(dc, ModuleType::get(module), TypeName,
-                                        lookupOptions)) {
-    for (auto pair : result) {
-      if (auto nominal = dyn_cast<NominalTypeDecl>(pair.Member)) {
-        cache = nominal->getDeclaredType();
-        return cache;
-      }
-    }
-  }
-
-  return nullptr;
-}
-
-Type TypeChecker::getNSObjectType(DeclContext *dc) {
-  return getObjectiveCNominalType(*this, NSObjectType, Context.Id_ObjectiveC,
-                                Context.getSwiftId(
-                                  KnownFoundationEntity::NSObject),
-                                dc);
-}
-
-Type TypeChecker::getNSErrorType(DeclContext *dc) {
-  return getObjectiveCNominalType(*this, NSErrorType, Context.Id_Foundation,
-                                  Context.getSwiftId(
-                                    KnownFoundationEntity::NSError),
-                                  dc);
-}
-
-Type TypeChecker::getObjCSelectorType(DeclContext *dc) {
-  return getObjectiveCNominalType(*this, ObjCSelectorType,
-                                  Context.Id_ObjectiveC,
-                                  Context.Id_Selector,
-                                  dc);
-}
-
 Type
 TypeChecker::getDynamicBridgedThroughObjCClass(DeclContext *dc,
                                                Type dynamicType,
@@ -3154,822 +3102,6 @@ Type TypeChecker::getSuperClassOf(Type type) {
     return ParenType::get(Context, superclassTy);
   }
   return type->getSuperclass();
-}
-
-static void lookupAndAddLibraryTypes(TypeChecker &TC,
-                                     ModuleDecl *Stdlib,
-                                     ArrayRef<Identifier> TypeNames,
-                                     llvm::DenseSet<CanType> &Types) {
-  SmallVector<ValueDecl *, 4> Results;
-  for (Identifier Id : TypeNames) {
-    Stdlib->lookupValue({}, Id, NLKind::UnqualifiedLookup, Results);
-    for (auto *VD : Results) {
-      if (auto *TD = dyn_cast<TypeDecl>(VD)) {
-        TC.validateDecl(TD);
-        Types.insert(TD->getDeclaredInterfaceType()->getCanonicalType());
-      }
-    }
-    Results.clear();
-  }
-}
-
-/// Emit an additional diagnostic describing why we are applying @objc to the
-/// decl, if this is not obvious from the decl itself.
-static void describeObjCReason(TypeChecker &TC, const ValueDecl *VD,
-                               ObjCReason Reason) {
-  if (Reason == ObjCReason::MemberOfObjCProtocol) {
-    VD->diagnose(diag::objc_inferring_on_objc_protocol_member);
-  } else if (Reason == ObjCReason::OverridesObjC) {
-    unsigned kind = isa<VarDecl>(VD) ? 0
-                  : isa<SubscriptDecl>(VD) ? 1
-                  : isa<ConstructorDecl>(VD) ? 2
-                  : 3;
-
-    auto overridden = VD->getOverriddenDecl();
-    overridden->diagnose(diag::objc_overriding_objc_decl,
-                         kind, VD->getOverriddenDecl()->getFullName());
-  } else if (Reason == ObjCReason::WitnessToObjC) {
-    auto requirement = TC.findWitnessedObjCRequirements(VD).front();
-    requirement->diagnose(diag::objc_witness_objc_requirement,
-                VD->getDescriptiveKind(), requirement->getFullName(),
-                cast<ProtocolDecl>(requirement->getDeclContext())
-                  ->getFullName());
-  }
-}
-
-static void diagnoseTypeNotRepresentableInObjC(const DeclContext *DC,
-                                               Type T,
-                                               SourceRange TypeRange) {
-  auto &diags = DC->getASTContext().Diags;
-
-  // Special diagnostic for tuples.
-  if (T->is<TupleType>()) {
-    if (T->isVoid())
-      diags.diagnose(TypeRange.Start, diag::not_objc_empty_tuple)
-          .highlight(TypeRange);
-    else
-      diags.diagnose(TypeRange.Start, diag::not_objc_tuple)
-          .highlight(TypeRange);
-    return;
-  }
-
-  // Special diagnostic for classes.
-  if (auto *CD = T->getClassOrBoundGenericClass()) {
-    if (!CD->isObjC())
-      diags.diagnose(TypeRange.Start, diag::not_objc_swift_class)
-          .highlight(TypeRange);
-    return;
-  }
-
-  // Special diagnostic for structs.
-  if (T->is<StructType>()) {
-    diags.diagnose(TypeRange.Start, diag::not_objc_swift_struct)
-        .highlight(TypeRange);
-    return;
-  }
-
-  // Special diagnostic for enums.
-  if (T->is<EnumType>()) {
-    diags.diagnose(TypeRange.Start, diag::not_objc_swift_enum)
-        .highlight(TypeRange);
-    return;
-  }
-
-  // Special diagnostic for protocols and protocol compositions.
-  if (T->isExistentialType()) {
-    if (T->isAny()) {
-      // Any is not @objc.
-      diags.diagnose(TypeRange.Start,
-                     diag::not_objc_empty_protocol_composition);
-      return;
-    }
-
-    auto layout = T->getExistentialLayout();
-
-    // See if the superclass is not @objc.
-    if (auto superclass = layout.explicitSuperclass) {
-      if (!superclass->getClassOrBoundGenericClass()->isObjC()) {
-        diags.diagnose(TypeRange.Start, diag::not_objc_class_constraint,
-                       superclass);
-        return;
-      }
-    }
-
-    // Find a protocol that is not @objc.
-    bool sawErrorProtocol = false;
-    for (auto P : layout.getProtocols()) {
-      auto *PD = P->getDecl();
-
-      if (PD->isSpecificProtocol(KnownProtocolKind::Error)) {
-        sawErrorProtocol = true;
-        break;
-      }
-
-      if (!PD->isObjC()) {
-        diags.diagnose(TypeRange.Start, diag::not_objc_protocol,
-                       PD->getDeclaredType());
-        return;
-      }
-    }
-
-    if (sawErrorProtocol) {
-      diags.diagnose(TypeRange.Start,
-                     diag::not_objc_error_protocol_composition);
-      return;
-    }
-
-    return;
-  }
-
-  if (T->is<ArchetypeType>() || T->isTypeParameter()) {
-    diags.diagnose(TypeRange.Start, diag::not_objc_generic_type_param)
-        .highlight(TypeRange);
-    return;
-  }
-
-  if (auto fnTy = T->getAs<FunctionType>()) {
-    if (fnTy->getExtInfo().throws() ) {
-      diags.diagnose(TypeRange.Start, diag::not_objc_function_type_throwing)
-        .highlight(TypeRange);
-      return;
-    }
-
-    diags.diagnose(TypeRange.Start, diag::not_objc_function_type_param)
-      .highlight(TypeRange);
-    return;
-  }
-}
-
-static void diagnoseFunctionParamNotRepresentable(
-    TypeChecker &TC, const AbstractFunctionDecl *AFD, unsigned NumParams,
-    unsigned ParamIndex, const ParamDecl *P, ObjCReason Reason) {
-  if (!shouldDiagnoseObjCReason(Reason, AFD->getASTContext()))
-    return;
-
-  if (NumParams == 1) {
-    AFD->diagnose(diag::objc_invalid_on_func_single_param_type,
-                  getObjCDiagnosticAttrKind(Reason));
-  } else {
-    AFD->diagnose(diag::objc_invalid_on_func_param_type,
-                  ParamIndex + 1, getObjCDiagnosticAttrKind(Reason));
-  }
-  if (P->hasType()) {
-    Type ParamTy = P->getType();
-    SourceRange SR;
-    if (auto typeRepr = P->getTypeLoc().getTypeRepr())
-      SR = typeRepr->getSourceRange();
-    diagnoseTypeNotRepresentableInObjC(AFD, ParamTy, SR);
-  }
-  describeObjCReason(TC, AFD, Reason);
-}
-
-static bool isParamListRepresentableInObjC(TypeChecker &TC,
-                                           const AbstractFunctionDecl *AFD,
-                                           const ParameterList *PL,
-                                           ObjCReason Reason) {
-  // If you change this function, you must add or modify a test in PrintAsObjC.
-
-  bool Diagnose = shouldDiagnoseObjCReason(Reason, TC.Context);
-
-  bool IsObjC = true;
-  unsigned NumParams = PL->size();
-  for (unsigned ParamIndex = 0; ParamIndex != NumParams; ParamIndex++) {
-    auto param = PL->get(ParamIndex);
-
-    // Swift Varargs are not representable in Objective-C.
-    if (param->isVariadic()) {
-      if (Diagnose && shouldDiagnoseObjCReason(Reason, TC.Context)) {
-        TC.diagnose(param->getStartLoc(), diag::objc_invalid_on_func_variadic,
-                    getObjCDiagnosticAttrKind(Reason))
-          .highlight(param->getSourceRange());
-        describeObjCReason(TC, AFD, Reason);
-      }
-
-      return false;
-    }
-
-    // Swift inout parameters are not representable in Objective-C.
-    if (param->isInOut()) {
-      if (Diagnose && shouldDiagnoseObjCReason(Reason, TC.Context)) {
-        TC.diagnose(param->getStartLoc(), diag::objc_invalid_on_func_inout,
-                    getObjCDiagnosticAttrKind(Reason))
-          .highlight(param->getSourceRange());
-        describeObjCReason(TC, AFD, Reason);
-      }
-
-      return false;
-    }
-
-    if (param->getType()->isRepresentableIn(
-          ForeignLanguage::ObjectiveC,
-          const_cast<AbstractFunctionDecl *>(AFD)))
-      continue;
-    
-    // Permit '()' when this method overrides a method with a
-    // foreign error convention that replaces NSErrorPointer with ()
-    // and this is the replaced parameter.
-    AbstractFunctionDecl *overridden;
-    if (param->getType()->isVoid() && AFD->hasThrows() &&
-        (overridden = AFD->getOverriddenDecl())) {
-      auto foreignError = overridden->getForeignErrorConvention();
-      if (foreignError &&
-          foreignError->isErrorParameterReplacedWithVoid() &&
-          foreignError->getErrorParameterIndex() == ParamIndex) {
-        continue;
-      }
-    }
-
-    IsObjC = false;
-    if (!Diagnose) {
-      // Save some work and return as soon as possible if we are not
-      // producing diagnostics.
-      return IsObjC;
-    }
-    diagnoseFunctionParamNotRepresentable(TC, AFD, NumParams, ParamIndex,
-                                          param, Reason);
-  }
-  return IsObjC;
-}
-
-/// Check whether the given declaration occurs within a constrained
-/// extension, or an extension of a class with generic ancestry, or an
-/// extension of an Objective-C runtime visible class, and
-/// therefore is not representable in Objective-C.
-static bool checkObjCInExtensionContext(TypeChecker &tc,
-                                        const ValueDecl *value,
-                                        bool diagnose) {
-  auto DC = value->getDeclContext();
-
-  if (auto ED = dyn_cast<ExtensionDecl>(DC)) {
-    if (ED->getTrailingWhereClause()) {
-      if (diagnose) {
-        tc.diagnose(value->getLoc(), diag::objc_in_extension_context);
-      }
-      return true;
-    }
-
-    // Check if any Swift classes in the inheritance hierarchy have generic
-    // parameters.
-    // FIXME: This is a current limitation, not inherent. We don't have
-    // a concrete class to attach Objective-C category metadata to.
-    if (auto generic = ED->getDeclaredInterfaceType()
-                           ->getGenericAncestor()) {
-      if (!generic->getClassOrBoundGenericClass()->hasClangNode()) {
-        if (diagnose) {
-          tc.diagnose(value, diag::objc_in_generic_extension);
-        }
-        return true;
-      }
-    }
-  }
-
-  return false;
-}
-
-/// Check whether the given declaration contains its own generic parameters,
-/// and therefore is not representable in Objective-C.
-static bool checkObjCWithGenericParams(TypeChecker &TC,
-                                       const AbstractFunctionDecl *AFD,
-                                       ObjCReason Reason) {
-  bool Diagnose = shouldDiagnoseObjCReason(Reason, TC.Context);
-
-  if (AFD->getGenericParams()) {
-    // Diagnose this problem, if asked to.
-    if (Diagnose) {
-      TC.diagnose(AFD->getLoc(), diag::objc_invalid_with_generic_params,
-                  getObjCDiagnosticAttrKind(Reason));
-      describeObjCReason(TC, AFD, Reason);
-    }
-
-    return true;
-  }
-
-  return false;
-}
-
-/// CF types cannot have @objc methods, because they don't have real class
-/// objects.
-static bool checkObjCInForeignClassContext(TypeChecker &TC,
-                                           const ValueDecl *VD,
-                                           ObjCReason Reason) {
-  bool Diagnose = shouldDiagnoseObjCReason(Reason, TC.Context);
-
-  auto type = VD->getDeclContext()->getDeclaredInterfaceType();
-  if (!type)
-    return false;
-
-  auto clas = type->getClassOrBoundGenericClass();
-  if (!clas)
-    return false;
-
-  switch (clas->getForeignClassKind()) {
-  case ClassDecl::ForeignKind::Normal:
-    return false;
-
-  case ClassDecl::ForeignKind::CFType:
-    if (Diagnose) {
-      TC.diagnose(VD, diag::objc_invalid_on_foreign_class,
-                  getObjCDiagnosticAttrKind(Reason));
-      describeObjCReason(TC, VD, Reason);
-    }
-    break;
-
-  case ClassDecl::ForeignKind::RuntimeOnly:
-    if (Diagnose) {
-      TC.diagnose(VD, diag::objc_in_objc_runtime_visible,
-                  VD->getDescriptiveKind(), getObjCDiagnosticAttrKind(Reason),
-                  clas->getName());
-      describeObjCReason(TC, VD, Reason);
-    }
-    break;
-  }
-
-  return true;
-}
-
-bool TypeChecker::isCIntegerType(const DeclContext *DC, Type T) {
-  if (CIntegerTypes.empty())
-    fillObjCRepresentableTypeCache(DC);
-  return CIntegerTypes.count(T->getCanonicalType());
-}
-
-/// Determines whether the given type is bridged to an Objective-C class type.
-static bool isBridgedToObjectiveCClass(DeclContext *dc, Type type) {
-  switch (type->getForeignRepresentableIn(ForeignLanguage::ObjectiveC, dc)
-            .first) {
-  case ForeignRepresentableKind::Trivial:
-  case ForeignRepresentableKind::None:
-    return false;
-
-  case ForeignRepresentableKind::Object:
-  case ForeignRepresentableKind::Bridged:
-  case ForeignRepresentableKind::BridgedError:
-  case ForeignRepresentableKind::StaticBridged:
-    return true;
-  }
-
-  llvm_unreachable("Unhandled ForeignRepresentableKind in switch.");
-}
-
-bool TypeChecker::isRepresentableInObjC(
-       const AbstractFunctionDecl *AFD,
-       ObjCReason Reason,
-       Optional<ForeignErrorConvention> &errorConvention) {
-  // Clear out the error convention. It will be added later if needed.
-  errorConvention = None;
-
-  // If you change this function, you must add or modify a test in PrintAsObjC.
-
-  bool Diagnose = shouldDiagnoseObjCReason(Reason, Context);
-
-  if (checkObjCInForeignClassContext(*this, AFD, Reason))
-    return false;
-  if (checkObjCWithGenericParams(*this, AFD, Reason))
-    return false;
-  if (checkObjCInExtensionContext(*this, AFD, Diagnose))
-    return false;
-
-  if (AFD->isOperator()) {
-    diagnose(AFD, (isa<ProtocolDecl>(AFD->getDeclContext())
-                   ? diag::objc_operator_proto
-                   : diag::objc_operator));
-    return false;
-  }
-
-  if (auto accessor = dyn_cast<AccessorDecl>(AFD)) {
-    // Accessors can only be @objc if the storage declaration is.
-    // Global computed properties may however @_cdecl their accessors.
-    auto storage = accessor->getStorage();
-    validateDecl(storage);
-    if (!storage->isObjC() && Reason != ObjCReason::ExplicitlyCDecl &&
-        Reason != ObjCReason::WitnessToObjC) {
-      if (Diagnose) {
-        auto error = accessor->isGetter()
-                  ? (isa<VarDecl>(storage) 
-                       ? diag::objc_getter_for_nonobjc_property
-                       : diag::objc_getter_for_nonobjc_subscript)
-                  : (isa<VarDecl>(storage)
-                       ? diag::objc_setter_for_nonobjc_property
-                       : diag::objc_setter_for_nonobjc_subscript);
-
-        diagnose(accessor->getLoc(), error);
-        describeObjCReason(*this, accessor, Reason);
-      }
-      return false;
-    }
-
-    switch (accessor->getAccessorKind()) {
-    case AccessorKind::DidSet:
-    case AccessorKind::WillSet:
-        // willSet/didSet implementations are never exposed to objc, they are
-        // always directly dispatched from the synthesized setter.
-      if (Diagnose) {
-        diagnose(accessor->getLoc(), diag::objc_observing_accessor);
-        describeObjCReason(*this, accessor, Reason);
-      }
-      return false;
-
-    case AccessorKind::Get:
-    case AccessorKind::Set:
-      return true;
-
-    case AccessorKind::MaterializeForSet:
-      // materializeForSet is synthesized, so never complain about it
-      return false;
-
-    case AccessorKind::Address:
-    case AccessorKind::MutableAddress:
-      if (Diagnose) {
-        diagnose(accessor->getLoc(), diag::objc_addressor);
-        describeObjCReason(*this, accessor, Reason);
-      }
-      return false;
-    }
-    llvm_unreachable("bad kind");
-  }
-
-  // As a special case, an initializer with a single, named parameter of type
-  // '()' is always representable in Objective-C. This allows us to cope with
-  // zero-parameter methods with selectors that are longer than "init". For
-  // example, this allows:
-  //
-  // \code
-  // class Foo {
-  //   @objc init(malice: ()) { } // selector is "initWithMalice"
-  // }
-  // \endcode
-  bool isSpecialInit = false;
-  if (auto init = dyn_cast<ConstructorDecl>(AFD))
-    isSpecialInit = init->isObjCZeroParameterWithLongSelector();
-
-  if (!isSpecialInit &&
-      !isParamListRepresentableInObjC(*this, AFD,
-                                      AFD->getParameterLists().back(),
-                                      Reason)) {
-    if (!Diagnose) {
-      // Return as soon as possible if we are not producing diagnostics.
-      return false;
-    }
-  }
-
-  if (auto FD = dyn_cast<FuncDecl>(AFD)) {
-    Type ResultType = FD->mapTypeIntoContext(FD->getResultInterfaceType());
-    if (!ResultType->hasError() &&
-        !ResultType->isVoid() &&
-        !ResultType->isUninhabited() &&
-        !ResultType->isRepresentableIn(ForeignLanguage::ObjectiveC,
-                                       const_cast<FuncDecl *>(FD))) {
-      if (Diagnose) {
-        diagnose(AFD->getLoc(), diag::objc_invalid_on_func_result_type,
-                 getObjCDiagnosticAttrKind(Reason));
-        SourceRange Range =
-            FD->getBodyResultTypeLoc().getTypeRepr()->getSourceRange();
-        diagnoseTypeNotRepresentableInObjC(FD, ResultType, Range);
-        describeObjCReason(*this, FD, Reason);
-      }
-      return false;
-    }
-  }
-
-  // Throwing functions must map to a particular error convention.
-  if (AFD->hasThrows()) {
-    DeclContext *dc = const_cast<AbstractFunctionDecl *>(AFD);
-    SourceLoc throwsLoc;
-    Type resultType;
-
-    const ConstructorDecl *ctor = nullptr;
-    if (auto func = dyn_cast<FuncDecl>(AFD)) {
-      resultType = func->getResultInterfaceType();
-      throwsLoc = func->getThrowsLoc();
-    } else {
-      ctor = cast<ConstructorDecl>(AFD);
-      throwsLoc = ctor->getThrowsLoc();
-    }
-
-    ForeignErrorConvention::Kind kind;
-    CanType errorResultType;
-    Type optOptionalType;
-    if (ctor) {
-      // Initializers always use the nil result convention.
-      kind = ForeignErrorConvention::NilResult;
-
-      // Only non-failing initializers can throw.
-      if (ctor->getFailability() != OTK_None) {
-        if (Diagnose) {
-          diagnose(AFD->getLoc(), diag::objc_invalid_on_failing_init,
-                   getObjCDiagnosticAttrKind(Reason))
-            .highlight(throwsLoc);
-          describeObjCReason(*this, AFD, Reason);
-        }
-
-        return false;
-      }
-    } else if (resultType->isVoid()) {
-      // Functions that return nothing (void) can be throwing; they indicate
-      // failure with a 'false' result.
-      kind = ForeignErrorConvention::ZeroResult;
-      NominalTypeDecl *boolDecl = Context.getObjCBoolDecl();
-      // On Linux, we might still run @objc tests even though there's
-      // no ObjectiveC Foundation, so use Swift.Bool instead of crapping
-      // out.
-      if (boolDecl == nullptr)
-        boolDecl = Context.getBoolDecl();
-
-      if (boolDecl == nullptr) {
-        diagnose(AFD->getLoc(), diag::broken_bool);
-        return false;
-      }
-
-      errorResultType = boolDecl->getDeclaredType()->getCanonicalType();
-    } else if (!resultType->getOptionalObjectType() &&
-               isBridgedToObjectiveCClass(dc, resultType)) {
-      // Functions that return a (non-optional) type bridged to Objective-C
-      // can be throwing; they indicate failure with a nil result.
-      kind = ForeignErrorConvention::NilResult;
-    } else if ((optOptionalType = resultType->getOptionalObjectType()) &&
-               isBridgedToObjectiveCClass(dc, optOptionalType)) {
-      // Cannot return an optional bridged type, because 'nil' is reserved
-      // to indicate failure. Call this out in a separate diagnostic.
-      if (Diagnose) {
-        diagnose(AFD->getLoc(),
-                 diag::objc_invalid_on_throwing_optional_result,
-                 getObjCDiagnosticAttrKind(Reason),
-                 resultType)
-          .highlight(throwsLoc);
-        describeObjCReason(*this, AFD, Reason);
-      }
-      return false;
-    } else {
-      // Other result types are not permitted.
-      if (Diagnose) {
-        diagnose(AFD->getLoc(),
-                 diag::objc_invalid_on_throwing_result,
-                 getObjCDiagnosticAttrKind(Reason),
-                 resultType)
-          .highlight(throwsLoc);
-        describeObjCReason(*this, AFD, Reason);
-      }
-      return false;
-    }
-
-    // The error type is always 'AutoreleasingUnsafeMutablePointer<NSError?>?'.
-    Type errorParameterType = getNSErrorType(dc);
-    if (errorParameterType) {
-      errorParameterType = OptionalType::get(errorParameterType);
-      errorParameterType
-        = BoundGenericType::get(
-            Context.getAutoreleasingUnsafeMutablePointerDecl(),
-            nullptr,
-            errorParameterType);
-      errorParameterType = OptionalType::get(errorParameterType);
-    }
-
-    // Determine the parameter index at which the error will go.
-    unsigned errorParameterIndex;
-    bool foundErrorParameterIndex = false;
-
-    // If there is an explicit @objc attribute with a name, look for
-    // the "error" selector piece.
-    if (auto objc = AFD->getAttrs().getAttribute<ObjCAttr>()) {
-      if (auto objcName = objc->getName()) {
-        auto selectorPieces = objcName->getSelectorPieces();
-        for (unsigned i = selectorPieces.size(); i > 0; --i) {
-          // If the selector piece is "error", this is the location of
-          // the error parameter.
-          auto piece = selectorPieces[i-1];
-          if (piece == Context.Id_error) {
-            errorParameterIndex = i-1;
-            foundErrorParameterIndex = true;
-            break;
-          }
-
-          // If the first selector piece ends with "Error", it's here.
-          if (i == 1 && camel_case::getLastWord(piece.str()) == "Error") {
-            errorParameterIndex = i-1;
-            foundErrorParameterIndex = true;
-            break;
-          }
-        }
-      }
-    }
-
-    // If the selector did not provide an index for the error, find
-    // the last parameter that is not a trailing closure.
-    if (!foundErrorParameterIndex) {
-      auto *paramList = AFD->getParameterLists().back();
-      errorParameterIndex = paramList->size();
-
-      // Note: the errorParameterIndex is actually a SIL function
-      // parameter index, which means tuples are exploded. Normally
-      // tuple types cannot be bridged to Objective-C, except for
-      // one special case -- a constructor with a single named parameter
-      // 'foo' of tuple type becomes a zero-argument selector named
-      // 'initFoo'.
-      if (auto *CD = dyn_cast<ConstructorDecl>(AFD))
-        if (CD->isObjCZeroParameterWithLongSelector())
-          errorParameterIndex--;
-
-      while (errorParameterIndex > 0) {
-        // Skip over trailing closures.
-        auto type = paramList->get(errorParameterIndex - 1)->getType();
-
-        // It can't be a trailing closure unless it has a specific form.
-        // Only consider the rvalue type.
-        type = type->getRValueType();
-        
-        // Look through one level of optionality.
-        if (auto objectType = type->getOptionalObjectType())
-          type = objectType;
-        
-        // Is it a function type?
-        if (!type->is<AnyFunctionType>()) break;
-        --errorParameterIndex;
-      }
-    }
-
-    // Form the error convention.
-    CanType canErrorParameterType;
-    if (errorParameterType)
-      canErrorParameterType = errorParameterType->getCanonicalType();
-    switch (kind) {
-    case ForeignErrorConvention::ZeroResult:
-      errorConvention = ForeignErrorConvention::getZeroResult(
-                          errorParameterIndex,
-                          ForeignErrorConvention::IsNotOwned,
-                          ForeignErrorConvention::IsNotReplaced,
-                          canErrorParameterType,
-                          errorResultType);
-      break;
-
-    case ForeignErrorConvention::NonZeroResult:
-      errorConvention = ForeignErrorConvention::getNonZeroResult(
-                          errorParameterIndex,
-                          ForeignErrorConvention::IsNotOwned,
-                          ForeignErrorConvention::IsNotReplaced,
-                          canErrorParameterType,
-                          errorResultType);
-      break;
-
-    case ForeignErrorConvention::ZeroPreservedResult:
-      errorConvention = ForeignErrorConvention::getZeroPreservedResult(
-                          errorParameterIndex,
-                          ForeignErrorConvention::IsNotOwned,
-                          ForeignErrorConvention::IsNotReplaced,
-                          canErrorParameterType);
-      break;
-
-    case ForeignErrorConvention::NilResult:
-      errorConvention = ForeignErrorConvention::getNilResult(
-                          errorParameterIndex,
-                          ForeignErrorConvention::IsNotOwned,
-                          ForeignErrorConvention::IsNotReplaced,
-                          canErrorParameterType);
-      break;
-
-    case ForeignErrorConvention::NonNilError:
-      errorConvention = ForeignErrorConvention::getNilResult(
-                          errorParameterIndex,
-                          ForeignErrorConvention::IsNotOwned,
-                          ForeignErrorConvention::IsNotReplaced,
-                          canErrorParameterType);
-      break;
-    }
-  }
-
-  return true;
-}
-
-bool TypeChecker::isRepresentableInObjC(const VarDecl *VD, ObjCReason Reason) {
-  // If you change this function, you must add or modify a test in PrintAsObjC.
-
-  if (VD->isInvalid())
-    return false;
-
-  Type T = VD->getDeclContext()->mapTypeIntoContext(VD->getInterfaceType());
-  if (auto *RST = T->getAs<ReferenceStorageType>()) {
-    // In-memory layout of @weak and @unowned does not correspond to anything
-    // in Objective-C, but this does not really matter here, since Objective-C
-    // uses getters and setters to operate on the property.
-    // Because of this, look through @weak and @unowned.
-    T = RST->getReferentType();
-  }
-  bool Result = T->isRepresentableIn(ForeignLanguage::ObjectiveC,
-                                     VD->getDeclContext());
-  bool Diagnose = shouldDiagnoseObjCReason(Reason, Context);
-
-  if (Result && checkObjCInExtensionContext(*this, VD, Diagnose))
-    return false;
-
-  if (checkObjCInForeignClassContext(*this, VD, Reason))
-    return false;
-
-  if (!Diagnose || Result)
-    return Result;
-
-  SourceRange TypeRange = VD->getTypeSourceRangeForDiagnostics();
-  // TypeRange can be invalid; e.g. '@objc let foo = SwiftType()'
-  if (TypeRange.isInvalid())
-    TypeRange = VD->getNameLoc();
-
-  diagnose(VD->getLoc(), diag::objc_invalid_on_var,
-           getObjCDiagnosticAttrKind(Reason))
-      .highlight(TypeRange);
-  diagnoseTypeNotRepresentableInObjC(VD->getDeclContext(),
-                                     VD->getInterfaceType(),
-                                     TypeRange);
-  describeObjCReason(*this, VD, Reason);
-
-  return Result;
-}
-
-bool TypeChecker::isRepresentableInObjC(const SubscriptDecl *SD,
-                                        ObjCReason Reason) {
-  // If you change this function, you must add or modify a test in PrintAsObjC.
-
-  bool Diagnose = shouldDiagnoseObjCReason(Reason, Context);
-
-  if (checkObjCInForeignClassContext(*this, SD, Reason))
-    return false;
-
-  // Figure out the type of the indices.
-  Type IndicesType = SD->getIndicesInterfaceType()->getWithoutImmediateLabel();
-  
-  if (IndicesType->hasError())
-    return false;
-
-  bool IndicesResult =
-    IndicesType->isRepresentableIn(ForeignLanguage::ObjectiveC,
-                                   SD->getDeclContext());
-
-  Type ElementType = SD->getElementInterfaceType();
-  bool ElementResult = ElementType->isRepresentableIn(
-        ForeignLanguage::ObjectiveC, SD->getDeclContext());
-  bool Result = IndicesResult && ElementResult;
-
-  if (Result && checkObjCInExtensionContext(*this, SD, Diagnose))
-    return false;
-
-  // Make sure we know how to map the selector appropriately.
-  if (Result && SD->getObjCSubscriptKind() == ObjCSubscriptKind::None) {
-    SourceRange IndexRange = SD->getIndices()->getSourceRange();
-    diagnose(SD->getLoc(), diag::objc_invalid_subscript_key_type,
-             getObjCDiagnosticAttrKind(Reason), IndicesType)
-      .highlight(IndexRange);
-    return false;
-  }
-
-  if (!Diagnose || Result)
-    return Result;
-
-  SourceRange TypeRange;
-  if (!IndicesResult)
-    TypeRange = SD->getIndices()->getSourceRange();
-  else
-    TypeRange = SD->getElementTypeLoc().getSourceRange();
-  diagnose(SD->getLoc(), diag::objc_invalid_on_subscript,
-           getObjCDiagnosticAttrKind(Reason))
-    .highlight(TypeRange);
-
-  diagnoseTypeNotRepresentableInObjC(SD->getDeclContext(),
-                                     !IndicesResult ? IndicesType
-                                                    : ElementType,
-                                     TypeRange);
-  describeObjCReason(*this, SD, Reason);
-
-  return Result;
-}
-
-bool TypeChecker::canBeRepresentedInObjC(const ValueDecl *decl) {
-  if (!Context.LangOpts.EnableObjCInterop)
-    return false;
-
-  if (auto func = dyn_cast<AbstractFunctionDecl>(decl)) {
-    Optional<ForeignErrorConvention> errorConvention;
-    return isRepresentableInObjC(func, ObjCReason::MemberOfObjCMembersClass,
-                                 errorConvention);
-  }
-
-  if (auto var = dyn_cast<VarDecl>(decl))
-    return isRepresentableInObjC(var, ObjCReason::MemberOfObjCMembersClass);
-
-  if (auto subscript = dyn_cast<SubscriptDecl>(decl))
-    return isRepresentableInObjC(subscript,
-                                 ObjCReason::MemberOfObjCMembersClass);
-
-  return false;
-}
-
-void TypeChecker::fillObjCRepresentableTypeCache(const DeclContext *DC) {
-  if (!CIntegerTypes.empty())
-    return;
-
-  SmallVector<Identifier, 32> StdlibTypeNames;
-  ModuleDecl *Stdlib = getStdlibModule(DC);
-#define MAP_BUILTIN_TYPE(_, __)
-#define MAP_BUILTIN_INTEGER_TYPE(CLANG_BUILTIN_KIND, SWIFT_TYPE_NAME) \
-  StdlibTypeNames.push_back(Context.getIdentifier(#SWIFT_TYPE_NAME));
-#include "swift/ClangImporter/BuiltinMappedTypes.def"
-  lookupAndAddLibraryTypes(*this, Stdlib, StdlibTypeNames, CIntegerTypes);
 }
 
 namespace {
