+++ lib/Sema/TypeChecker.h
@@ -570,99 +570,7 @@ withoutContext(TypeResolutionOptions options, bool preserveSIL = false) {
   return options;
 }
 
-/// Describes the reason why are we trying to apply @objc to a declaration.
-///
-/// Should only affect diagnostics. If you change this enum, also change
-/// the OBJC_ATTR_SELECT macro in DiagnosticsSema.def.
-enum class ObjCReason {
-  /// Has the '@cdecl' attribute.
-  ExplicitlyCDecl,
-  /// Has the 'dynamic' modifier.
-  ExplicitlyDynamic,
-  /// Has an explicit '@objc' attribute.
-  ExplicitlyObjC,
-  /// Has an explicit '@IBOutlet' attribute.
-  ExplicitlyIBOutlet,
-  /// Has an explicit '@IBAction' attribute.
-  ExplicitlyIBAction,
-  /// Has an explicit '@NSManaged' attribute.
-  ExplicitlyNSManaged,
-  /// Is a member of an @objc protocol.
-  MemberOfObjCProtocol,
-  /// Implicitly-introduced @objc.
-  ImplicitlyObjC,
-  /// Is an override of an @objc member.
-  OverridesObjC,
-  /// Is a witness to an @objc protocol requirement.
-  WitnessToObjC,
-  /// Has an explicit '@IBInspectable' attribute.
-  ExplicitlyIBInspectable,
-  /// Has an explicit '@GKInspectable' attribute.
-  ExplicitlyGKInspectable,
-  /// Is it a member of an @objc extension of a class.
-  MemberOfObjCExtension,
-  /// Is it a member of an @objcMembers class.
-  MemberOfObjCMembersClass,
-  /// A member of an Objective-C-defined class or subclass.
-  MemberOfObjCSubclass,
-  /// An accessor to a property.
-  Accessor,
-};
-
-/// Determine whether we should diagnose conflicts due to inferring @objc
-/// with this particular reason.
-static inline bool shouldDiagnoseObjCReason(ObjCReason reason,
-                                            ASTContext &ctx) {
-  switch(reason) {
-  case ObjCReason::ExplicitlyCDecl:
-  case ObjCReason::ExplicitlyDynamic:
-  case ObjCReason::ExplicitlyObjC:
-  case ObjCReason::ExplicitlyIBOutlet:
-  case ObjCReason::ExplicitlyIBAction:
-  case ObjCReason::ExplicitlyNSManaged:
-  case ObjCReason::MemberOfObjCProtocol:
-  case ObjCReason::OverridesObjC:
-  case ObjCReason::WitnessToObjC:
-  case ObjCReason::ImplicitlyObjC:
-  case ObjCReason::MemberOfObjCExtension:
-    return true;
-
-  case ObjCReason::ExplicitlyIBInspectable:
-  case ObjCReason::ExplicitlyGKInspectable:
-    return !ctx.LangOpts.EnableSwift3ObjCInference;
-
-  case ObjCReason::MemberOfObjCSubclass:
-  case ObjCReason::MemberOfObjCMembersClass:
-  case ObjCReason::Accessor:
-    return false;
-  }
-}
-
-/// Return the %select discriminator for the OBJC_ATTR_SELECT macro used to
-/// complain about the correct attribute during @objc inference.
-static inline unsigned getObjCDiagnosticAttrKind(ObjCReason reason) {
-  switch (reason) {
-  case ObjCReason::ExplicitlyCDecl:
-  case ObjCReason::ExplicitlyDynamic:
-  case ObjCReason::ExplicitlyObjC:
-  case ObjCReason::ExplicitlyIBOutlet:
-  case ObjCReason::ExplicitlyIBAction:
-  case ObjCReason::ExplicitlyNSManaged:
-  case ObjCReason::MemberOfObjCProtocol:
-  case ObjCReason::OverridesObjC:
-  case ObjCReason::WitnessToObjC:
-  case ObjCReason::ImplicitlyObjC:
-  case ObjCReason::ExplicitlyIBInspectable:
-  case ObjCReason::ExplicitlyGKInspectable:
-  case ObjCReason::MemberOfObjCExtension:
-    return static_cast<unsigned>(reason);
-
-  case ObjCReason::MemberOfObjCSubclass:
-  case ObjCReason::MemberOfObjCMembersClass:
-  case ObjCReason::Accessor:
-    llvm_unreachable("should not diagnose this @objc reason");
-  }
-}
+enum class ObjCReason;
 
 /// Flags that control protocol conformance checking.
 enum class ConformanceCheckFlags {
